Add optional SNS parameter
Added an optional parameter to specify the SNS topic ARN directly, for those pipelines that do not export the ARN. Added similar branching logic for configuring the event. Additionally fixed a problem with the example SSM commands. When populating the URL for the webhook, the cli will try to resolve it. Only way around this is to use json.