Bugfix email handler
The changes in #3 refactored handling of dependencies to make it more testable. It appears I did not fully refactor the email handler, in particular for how it instantiates the AWS.CodePipeline dependency, which caused issues with it putting the approvals. Changed the handler to properly instantiate the CodePipeline and fixed the spies in the specs to match.